Monthly Overview Stats Functionality
Stats for each day of the month are now able to be viewed on the calendar. Users
may also view previous months progress as well. Possible added functionality
later down the road would to be to show all days on the calendar(those left over
days from the month prior and the first few days of the next month).